requirements.txt
django
djangorestframework
plotly
psutil
requests
orca